pyproject.toml
[project]
name = "lab14-ai-evaluation-benchmarking"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.96.0",
    "chromadb>=1.5.8",
    "faiss-cpu>=1.13.2",
    "flashrank>=0.2.10",
    "jinja2>=3.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.15",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.2",
    "openai>=1.10.0",
    "pandas>=2.1.0",
    "pdf2image>=1.17.0",
    "pinecone-client>=6.0.0",
    "pypdf>=4.0.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.0",
    "ragas>=0.1.7",
    "rank-bm25>=0.2.2",
    "streamlit>=1.56.0",
    "tqdm>=4.66.0",
    "unstructured>=0.22.22",
]
